+++ Kernel/System/MailAccount/POP3TLS.pm
@@ -11,13 +11,22 @@ package Kernel::System::MailAccount::POP3TLS;
 use strict;
 use warnings;
 
-use Net::SSLGlue::POP3;
+use Net::POP3;
 
 use base qw(Kernel::System::MailAccount::POP3);
 
 our @ObjectDependencies = (
     'Kernel::System::Log',
 );
+
+# Use Net::SSLGlue::POP3 on systems with older Net::POP3 modules that cannot handle POP3S.
+BEGIN {
+    if ( !defined &Net::POP3::starttls ) {
+        ## nofilter(TidyAll::Plugin::OTRS::Perl::Require)
+        ## nofilter(TidyAll::Plugin::OTRS::Perl::SyntaxCheck)
+        require Net::SSLGlue::POP3;
+    }
+}
 
 sub Connect {
     my ( $Self, %Param ) = @_;